feat(client): cache is supported in web/browser platform
In web/browser platform, Cache is supported by `localStorage`.

Breaking change: `InMemoryCache` no longer accept `storageProvider`,
however this param has always been using for testing only.

The naming `InMemoryCache` is also debatable, as both `io` and `html`
implementation of this class is based on "persistent storage", not at all
"in memory".

There are refactor works to do to consolidate `in_memory_*` files.
