+++ packages/graphql/lib/src/cache/normalized_in_memory.dart
@@ -23,8 +23,7 @@ class NormalizedInMemoryCache extends InMemoryCache {
   NormalizedInMemoryCache({
     @required this.dataIdFromObject,
     this.prefix = '@cache/reference',
-    @required StorageProvider storageProvider,
-  }) : super(storageProvider: storageProvider);
+  });
 
   DataIdFromObject dataIdFromObject;
