+++ packages/graphql/lib/src/cache/optimistic.dart
@@ -5,7 +5,6 @@ import 'package:meta/meta.dart';
 import 'package:graphql/src/utilities/helpers.dart';
 import 'package:graphql/src/cache/cache.dart';
 import 'package:graphql/src/cache/normalized_in_memory.dart';
-import 'package:graphql/src/cache/in_memory.dart' show StorageProvider;
 import 'package:graphql/src/cache/lazy_cache_map.dart';
 
 class OptimisticPatch extends Object {
@@ -49,7 +48,7 @@ class OptimisticProxy implements Cache {
 
   // TODO should persistence be a seperate concern from caching
   @override
-  void save() {}
+  Future<void> save() async {}
   @override
   void restore() {}
   @override
@@ -62,11 +61,9 @@ class OptimisticCache extends NormalizedInMemoryCache {
   OptimisticCache({
     @required DataIdFromObject dataIdFromObject,
     String prefix = '@cache/reference',
-    @required StorageProvider storageProvider,
   }) : super(
           dataIdFromObject: dataIdFromObject,
           prefix: prefix,
-          storageProvider: storageProvider,
         );
 
   @protected
