+++ packages/graphql/test/in_memory_storage_test.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io' show Directory, FileSystemException;
 import 'package:test/test.dart';
 import 'package:graphql/src/cache/in_memory.dart';
 
@@ -39,29 +38,20 @@ final Map<String, Object> eData = <String, Object>{
   }
 };
 
-final Future<Directory> customStorageDirectory =
-    Directory.systemTemp.createTemp('file_test_');
-
 void main() {
   group('Normalizes writes', () {
     test('.write .read round trip', () async {
-      final InMemoryCache cache = InMemoryCache(
-        storageProvider: () => customStorageDirectory,
-      );
+      final InMemoryCache cache = InMemoryCache();
       cache.write(aKey, aData);
       await cache.save();
       cache.reset();
       await cache.restore();
       expect(cache.read(aKey), equals(aData));
-    }, onPlatform: {
-      "browser": Skip("Browser does not support dart:io (see #295)")
     });
 
     test('.write avoids overriding a superset with a subset of a field (#155)',
         () async {
-      final InMemoryCache cache = InMemoryCache(
-        storageProvider: () => customStorageDirectory,
-      );
+      final InMemoryCache cache = InMemoryCache();
       cache.write(aKey, aData);
 
       final Map<String, Object> anotherAData = <String, Object>{
@@ -80,14 +70,10 @@ void main() {
           'a': {'__typename': 'A', 'key': 'val'}
         }),
       );
-    }, onPlatform: {
-      "browser": Skip("Browser does not support dart:io (see #295)")
     });
 
     test('.write does not mutate input', () async {
-      final InMemoryCache cache = InMemoryCache(
-        storageProvider: () => customStorageDirectory,
-      );
+      final InMemoryCache cache = InMemoryCache();
       cache.write(aKey, aData);
       final Map<String, Object> anotherAData = <String, Object>{
         'a': <String, Object>{
@@ -108,14 +94,10 @@ void main() {
           'a': {'key': 'val'}
         }),
       );
-    }, onPlatform: {
-      "browser": Skip("Browser does not support dart:io (see #295)")
     });
 
     test('saving concurrently wont error', () async {
-      final InMemoryCache cache = InMemoryCache(
-        storageProvider: () => customStorageDirectory,
-      );
+      final InMemoryCache cache = InMemoryCache();
       cache.write(aKey, aData);
       cache.write(bKey, bData);
       cache.write(cKey, cData);
@@ -138,31 +120,6 @@ void main() {
       expect(cache.read(cKey), equals(cData));
       expect(cache.read(dKey), equals(dData));
       expect(cache.read(eKey), equals(eData));
-    }, onPlatform: {
-      "browser": Skip("Browser does not support dart:io (see #295)")
     });
   });
-
-  group('Exception handling', () {
-    test('FileSystemException', overridePrint((List<String> log) async {
-      final InMemoryCache cache = InMemoryCache(
-        storageProvider: () {
-          throw FileSystemException();
-        },
-      );
-      await cache.restore();
-      expect(
-        log,
-        ['Can\'t read file from storage, returning an empty HashMap.'],
-      );
-    }));
-  });
 }
-
-overridePrint(testFn(List<String> log)) => () {
-      final log = <String>[];
-      final spec = new ZoneSpecification(print: (_, __, ___, String msg) {
-        log.add(msg);
-      });
-      return Zone.current.fork(specification: spec).run(() => testFn(log));
-    };
