+++ packages/graphql/test/normalized_in_memory_test.dart
@@ -1,5 +1,3 @@
-import 'dart:io' show Directory;
-
 import 'package:test/test.dart';
 
 import 'package:graphql/src/cache/normalized_in_memory.dart';
@@ -189,7 +187,6 @@ final Map<String, Object> cyclicalObjNormalizedB = {
 
 NormalizedInMemoryCache getTestCache() => NormalizedInMemoryCache(
       dataIdFromObject: typenameDataIdFromObject,
-      storageProvider: () => Directory.systemTemp.createTemp('file_test_'),
     );
 
 void main() {
